[Test gap] Fix bug in test_disk_exhaustion.py  (#6855)
What is the motivation for this PR?
After exhausting disk space with fallocate command, we try to test SSH connect and IP packet forward, but if any exception is raised among that, this module would fail, and post steps outside this TC would be executed. But without releasing the disk space, it would cause other errors like AnsibleConnectionFailure: Host unreachable. So add the final block to wait the space to be released in order to not impact post steps.

How did you do it?
Execute the test behavior in the try block, and wait for disk space to be released in the final block.

How did you verify/test it?
Run TC and TC passed or failed in expected result.

Signed-off-by: Chun'ang Li <chunangli@microsoft.com>