Adding and modifying files for VxLAN-CRM testcases.
The following tests are covered. The NHs are not counted, so they
are done differently.

1. CRM cases are  pulled out of the main script, and will be kept in a seperate script.
- This is needed because the number of routes for CRM is different from the main script.
- The BFD scale will not work with the CRM, since the ptf can’t handle that many number of monitors.

The first one has 2 parts, one is number of routes and the other is number of nexthops.
The nexthops count increments only if nexthop really changes. Since we use default route for all our
traffic, this number doesn’t go up during the script execution.  So we use a different method(using
arp entries) to check this.