Enhancing core_dump_and_config_check to be multi-asic aware (#6527)
What is the motivation for this PR?
In our pipeline runs, autorestart tests against a multi-asic DUT were failing with error '
Feature 'x' auto-restart is not consistent across namespaces
The reason was that pretest goes and disables autorestart on all the containers. Then ACL tests run with change the config_db's and to retore in it's cleanup does a config load_minigraph. The above results in setting the autorestart state of the containers back to default of 'enabled'
Now, when check_dut_health_status kicks in, it takes a snapshot of the config_db.json before the ACL suite which has the autorestart state as 'disabled'. But, after the ACL suite, it detects that autorestart state has changed to 'enabled'. Thus, it tries to restore it.
However, when it restores, it only restores config_db.json, and not the other asics config_db's.
This results in the state of autorestart to be not consistent across namespaces - global has autorestart 'disabled', while namespace has autorestart 'enabled'.
How did you do it?
Fix for the above it to have enhance check_dut_health_status to be multi-asic aware.
It compares not just config_db.json, but also the config_db's of all the asics.
If it finds that config has changed, restore not just config_db.json, but also the config_db's of all the asics before rebooting the DUT.
How did you verify/test it?
Ran check_dut_health_status with changes to config_db's and validated that it is restored to what it is before the suite is run