[LogAnalyzer]Add ignore expected loganalyzer exceptions for techsupport
What is the motivation for this PR?
In Mellanox, when techsupport is taken, it invokes fw dump. While taking
the fw dump, the fw is busy and doesn't respond to other calls. The
access of sfp eeprom happens through firmware and xcvrd gets the DOM
fields every 60 seconds which fails during the fw dump.
This is a temporary issue and this log can be ignored.

How did you do it?
Add ignore expected loganalyzer exceptions for techsupport

How did you verify/test it?
Run techsupport testscript

Any platform specific information?
Supported testbed topology if it's a new test case?