Merged PR 1322: Split Test Pipeline for PTF and T0 cases on sai 1.7 and sai 1.10
Task Link: [Here](https://msazure.visualstudio.com/One/_workitems/edit/16340675)

For doing that, we can reuse the existing code for sai 1.7.
After enable the sai 1.10 pipeline, we need to re-organize the file strucure. move the currernt sai pipeline template and instance to a new folder for sai_qualify

For testing:
[SAI_v1.7_t0_nightly, SAI_v1.7_ptf_nightly, SAI_v1.10_ptf_nightly](https://dev.azure.com/mssonic/internal/_build?definitionScope=%5CDraft%5Czhiyuanliang)