Moving check for reboot cause after interface status check (#6721)
Currently if there is an issue in reboot cause, it immediately aborts the test case and moves on to the next script. Because interface/transceivers check is not done and the timeout for it not exercised, it can cause other scripts to prematurely error out stating that the interfaces are not up.