Testcase to verify that lossless traffic is not dropped during congesion. (#6853)
What is the motivation for this PR?
The test gap was identified during testing. The losslless traffic should not have any drops during the over-subscription of any port.

How did you do it?
Added 2 files for the test and the helper file.

How did you verify/test it?
Ran the script 3 times, all passed. Run time was 220 seconds.