Set worker numbers using variables in azure pipeline library (#6927)
Approach
What is the motivation for this PR?
Previously, we hard code the min and max numbers of instance in a plan. In this pr, we support passing the instance numbers of a testplan.

How did you do it?
Use a variable to set the instance number.

How did you verify/test it?
Any platform specific information?
Supported testbed topology if it's a new test case?