Update the fwutil function to support define the component for specific setup (#6756)
In the existing fwutil test implement, user can only define the component(BIOS, ONIE, CPLD) based on the platform type, if for the same platform, it require to define different components for different dut(such as some setup are respined), them the origin implementation dose not support. modify the script to support such scenario.
The fwutil test case should not be skipped, since the [test_fwutil]fixture 'fw_pkg_name' not found #6489 is not a real issue.
Fix some pep8 issue