[advanced-reboot] Handle logs in tmpfs: backup two log files before reboot (#6923)
Advanced reboot and upgrade path testcases sometimes fail when LogAnalyzer start string is missing in the logs after reboot. This issue happens on small disk devices, where logs are stored in tmpfs, and are lost after reboot.

The present case already takes a backup of logs before reboot. However backing up last log file is sometimes not sufficient, and log rotate can get triggered, and the log with LogAnalyzer string is no more in the latest log file.

In this PR, backup is done for last two log files (instead of one). This is applied to all the types of logs - syslog, sairedis, bgp logs.

How did you verify/test it?
Tested on a physical testbed with logs in tmpfs.