Remove timeout in each step. (#6947)
Description of PR
Previously, we set timeout in each step such as Lock testbed, Prepare testbed, Run test and KVM dump. When some issue suck like retry happens in one step, it will cause timeout error, but actually, it only needs more time to success. In this pr, we remove the timeout limit in each step and control the timeout outside in each job. When the job runs more than four hours, it will be cancelled.

What is the motivation for this PR?
Previously, we set timeout in each step such as Lock testbed, Prepare testbed, Run test and KVM dump. When some issue suck like retry happens in one step, it will cause timeout error, but actually, it only needs more time to success. In this pr, we remove the timeout limit in each step and control the timeout outside in each job. When the job runs more than four hours, it will be cancelled.

How did you do it?
Remove the timeout parameter in each step, and control the timeout outside in each job.
Set the default value of timeout to -1
How did you verify/test it?
Set the timeout of one job to 4 hours, and when timeout happens, azure pipeline will cancel this job.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>