Reboot dict in Inventory file for platform specific wait and timeout (#6719)
Current reboot logic has a flat timeout and is more catered towards fixed chassis which has a faster reboot time. This PR is to accommodate distributed chassis and to allow individual platforms to define their own wait and timeout in inventory file.

timeout and wait time can differ under different conditions, so added support to check if there is a testcase/class level definition for them.