[buffer] To save execution time, optimize test_buffer_deployment (#6985)
The buffer config, and most of time is spent fetching data from dut. So, the bigger the total number of port is, the longer the execution time is.
Since during the time of test, there is no config and status change, so we can fetch all db info one time for all ports instead of fetching info more times for one port. So, It can save much time. How much time is saved depends on the total number of port. The bigger the total number of port is, the longer the saved time is. Take sn3800 with 70 ports as an example: previously about 22 minutes, now about 2 minutes