[ansible] make deployment j2cli installation safe
When no j2cli executable can be detected, the deployment script
ansible/setup-management-network.sh tries to install it using pip.

According to the documentation at docs/testbed/README.testbed.Setup.md
this script must be launched with sudo during the deployment process.

Running "sudo pip install" can be OK in a CI or inside a test container
but it generates a clear warning.
It is strongly disadvised to run "pip install" under root in a user or
a production environment because it can break the host packages system.
"pip install --user" or venv are usually prefered in most situations.

sudo -H / --set-home option can ensure the HOME variable is set to the
user home directory so that "pip install --user" called from a script
with sudo does not use /root.
This way, installed executables remains available to the user w/o sudo.

Signed-off-by: Guillaume Lambert <guillaume.lambert@orange.com>