[Mellanox]Increase the wait time in crm nexthop and neighbor test
Sometimes the nexthop/neighbor check in crm test will fail because the available counter is not restored to the original value after the nexthop/neighbor is removed.
The root cause is when the nexthop/neighbor is removed there will be 2 steps to fully release the resources.
First, almost immediately, the resource occupied by neighbor and nexthop will be released.
And then after some time（may be more than current 30s on some platforms）, the resource occupied by the SDK/SAI internal object will be cleaned up. Only after that, the available count will restore to the original value.
Increase the wait_until timeout to 60s to make sure the internal object is cleaned.