Merged PR 1505: Add a tool for managing nightly test agents
Add a tool for managing nightly test agents

The nightly test agents of starlab are managed by K8S
running on trusty9 which will be deprecated.
Nightly test agents of SVC and BJW lab are managed
manually. Because of these, I built this tool for managing
the nightly test agents.

Documentation: `agent-manager.md`.
Signed-off-by: Xin Wang <xiwang5@microsoft.com>