[chassis] fix error with sonic-clear commands (#7180)
What is the motivation for this PR?
The tests in the test_drop_counter.py fail on the T2 testbed with both single and multi asic linecards.
In base_verification the rif counters are cleared for all dut for a given asic_index. In a test bed with mix of single and multi asic linecards the asic_index may not valid on all the linecards. This causes the sonic clear commands to fail and test exists.

Similar problem is present in the get_pkt_drops function

How did you do it?
in base_verification clear the counters for all the namespaces on all the duts in the testbed
in the get_pkt_drops, add check to add the namespace command suffix/prefix only when the namespace is valid.

Signed-off-by: Arvindsrinivasan Lakshmi Narasimhan <arlakshm@microsoft.com>