Revert "[testbed] Convert port alias to name when generating connection graph (#7110)" (#7199)
Reverts #7110 
I found there is a bug in #7110. From port_utils.py we can see many platforms are using Ethernet* as both port name and alias (for example, Ethernet1 appears in both port names and alias on platform Arista-7050-QX-32S). So, it's hard to tell if Ethernet* in link.csv is a name or an alias. I'll open a new PR later to take this situation into consideration.
Signed-off-by: Zhijian Li <zhijianli@microsoft.com>