Merged PR 1547: Minor fix for platform_tests branch aggregation
In PR https://dev.azure.com/mssonic/internal/_git/sonic-mgmt-int/pullrequest/1541,
we aggregate branch for platform_tests, will not report duplicated case if it has reported before for other branches.

In this PR, fix a minor bug in duplication check logic, early break if it turns out to be similar duplicated case.

For platform_tests, if case title is:
[platform_tests.api.test_chassis.TestChassisApi][test_get_watchdog][**master**][broadcom][Celestica-DX010-C32]
there is active IcM:
[platform_tests.api.test_chassis.TestChassisApi][test_get_watchdog][**internal**][broadcom][Celestica-DX010-C32]
analyzer considers them are duplicated, will not report this case.

Signed-off-by: Zhaohui Sun <zhaohuisun@microsoft.com>