[chassis] Fix for critical service check failing post config reload. (#7224)
After PR: #7003 for chassis/t2 we added macsec as critical service.
However macsec will not run on all LC's but only on platform where macsec_enabled=1 mostly the uplink LC's.

How I did:
Based on Feature table state update critical service list. Previously we were doing this only for supervisor but now we will do for LC also. We can do the check for all devices but it seems adding this check increases run-time #6873 so doing only for chassis as of now.