Merged PR 1555: Include all testbeds on server_61 for deep-clean
Include all testbeds on server_61 for deep-clean

Some new testbeds were added to server_61. However, they
are not included in the deep_clean job which runs on every
Saturday. After deep clean, the job is not able to bring up
testbeds not included in the deep clean pipeline yaml.

This change added the new testbeds to the deep clean
pipeline yaml to ensure that the testbeds are recovered
after deep clean.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>