[TestbedV2]Remove parameter expected_states in `.azure-pipelines/test_plan.py` (#7215)
Description of PR
In pr #6837, it improves the display of pipeline. But it still has duplicate parameter expected_states in order to compate to sonic-buildimage repo. After the modification in sonic-buildimage repo, I remove this parameter.

What is the motivation for this PR?
In pr #6837, it improves the display of pipeline. But it still has duplicate parameter expected_states in order to compate to sonic-buildimage repo. After the modification in sonic-buildimage repo, I remove this parameter.

How did you do it?
Remove the patameter expected_states in .azure-pipelines/test_plan.py and the code compating to sonic-buildimage.

How did you verify/test it?
By running pipeline.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>