[tacacs] Fix tacacs test remove default server in teardown (#7232)
What is the motivation for this PR?
According to the current logic of remove_all_tacacs_server, if the dut is not configured with default tacacs_server, it will cause an error in the find_server_command. If the dut is configured with default tacacs_server, the default tacacs_server will be deleted by mistake.

How did you do it?
Add module_ignore_errors for cmd finding tacacs server.
Make sure tacacs server to be deleted is not default tacacs server.
How did you verify/test it?
Run tacacs test on mx, m0, t0, t1 testbed. All passed.

Signed-off-by: Yaqiang Zhu <yaqiangzhu@microsoft.com>