Merged PR 1597: Improve deep clean to avoid hard coded servers and testbeds
Improve deep clean to avoid hard coded servers and testbeds

The deep clean pipeline needs a dict of servers and testbeds to run deep
clean. To avoid hard code servers and testbeds information in the deep
clean job, we can use this job to dynamically get the current server and
testbed information from testbed.yaml file. Then use Azure DevOps API
to trigger the deep clean pipeline with the collected server and testbed
info as parameter.

Using this method, we no long need to hard code the available servers
and testbeds in the deep clean pipeline yaml.

This trigger pipeline supports specifying the servers and testbeds to
skip deep clean. If just a server is specified in parameter, skip deep
clean on the whole server. If a server has testbed specified, deep
clean still runs for the server. The specified testbed will not be
recovered after deep clean.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>