Skip checking LineCard that are already on Inventory file Skip module list to take care logical chassis scenario (#7295)
n our testbed where we share a physical chassis that uses the same Supervisor card but group different LCs for different "logical chassis" so for some "logical chassis" some LCs are added to "skip_mod_list" as they are not meant to be tested for that logical chassis. Our expectation is that if it is marked as skipped, it is not meant for testing and no state of those skipped modules should be used for any validation logic. We should always trust what the inventory marked as skipped and not try to look at whatever state the "skipped module" should be.
In this particular testcase it is not handling the "logical chassis" scenario correctly and treating the linecards on the skip list must always have "Reachability" being "False". This logic causes logical testbed to always fail the test.