Fix skip multi asic error (#7243)
Description of PR
In pr #7204, it skips test_snmp_queue on multi-asic vs checks. And the condition branch in ['202205'] seems to be false, because in the test log, the branch name is 202205-12722. So in this pr, it changes this condition.

What is the motivation for this PR?
In pr #7204, it skips test_snmp_queue on multi-asic vs checks. And the condition branch in ['202205'] seems to be false, because in the test log, the branch name is 202205-12722. So in this pr, it changes this condition.

How did you do it?
Change the condition here.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>