[dualtor][active-active] Fix loganalyzer connection refused error
In loganalyzer fixture, the markers in the fixture setup is returned by
parallel_run is a proxy dict object. It has problem connecting to the
proxy manager server sometimes as the proxy manager is already deleted
as last call of parallel_run is returned.

So let's use built-in dict object as return for parallel_run.

Signed-off-by: Longxiang Lyu <lolv@microsoft.com>