[Mellanox]Increase the wait time in crm nexthop and neighbor test (#7112)
Sometimes the nexthop/neighbor check in crm test will fail on Mellanox paltforms because the available counter is not restored to the original value after the nexthop/neighbor is removed.
The root cause is when the nexthop/neighbor is removed there will be 2 steps to fully release the resources. First, almost immediately, the resource occupied by neighbor and nexthop will be released. And then after some time（may be more than current 30s on some platforms）, the resource occupied by the SDK/SAI internal object will be cleaned up. Only after that, the available count will restore to the original value. Increase the wait_until timeout to 60s to make sure the internal object is cleaned.

- What is the motivation for this PR?
Sometimes the nexthop/neighbor check in crm test will fail on Mellanox paltforms because the available counter is not restored to the original value after the nexthop/neighbor is removed. This PR is to fix this issue.

- How did you do it?
Increase the wait time to 60s for nexthop and neighbor check.

- How did you verify/test it?
By automation, all neighbor and nexthop cases passed.