[dualtor] Improve toggle fixtures (#7362)
Approach
What is the motivation for this PR?
The toggle utilities fail to toggle the mux ports in nightlies.
It is because that, without icmp_responder, linkmgrd is stuck in wait state and it has larger mux probe interval. So the mux status sync between linkmgrd and the mux simulator takes longer time.

Signed-off-by: Longxiang Lyu lolv@microsoft.com

How did you do it?
Add the following improvements:

add wait for mux toggle success
proactively probe the mux for unchanged ports after toggle to reduce the wait time
How did you verify/test it?