Skip test_trap_config_save_after_reboot for 202012 branch (#7373)
What is the motivation for this PR?
test_trap_config_save_after_reboot case should skipped in 202012 branch, since always_enable feature is not merged into 202012 image.
[CoPP] Add always_enabled field to coppmgr logic by noaOrMlnx · Pull Request #2034 · sonic-net/sonic-swss (github.com)

Test case was added in this PR:

xfail in this PR
[test] skip/xfail unmerged copp feature temporarily by yejianquan · Pull Request #5148 · sonic-net/sonic-mgmt (github.com)

How did you do it?
Xfail is not correct, it needs to be skipped for now.
Change xfail to skip.
Signed-off-by: Zhaohui Sun <zhaohuisun@microsoft.com>