Headroom Check Test Case - PFC Test Gap (#6924)
Description of PR
Summary:

This PR is to introduce two new test cases to test headroom capacities for pfc enabled traffic. For now, all the new test cases are conditioned for Cisco 8102 and Arista 7050CX3.

No.	Test case	What to verify
1	test_pfc_single_lossless_headroom	The test case is to test headroom capacity for DUT for a single lossless priority.
2	test_pfc_pause_multi_lossless_headroom	The test case is to test headroom capacity for DUT for multiple lossless priorities.

Both tests check for cases when headroom should not be exceeded, and when it should, and thus fail.