[voq][chassis] Ingoring local asic ping of ipv6 interfaces when one side is down (#6318)
What is the motivation for this PR?
Local asic ping of ipv6 interfaces fails when one side is down. We get the error message "ping: bind icmp socket: Cannot assign requested address"

How did you do it?
We ignore local asic ping of ipv6 interfaces if version is not v4.