Remove try-except block from fixture `fanouthosts` (#7421)
What is the motivation for this PR?
The try-except block in fixture fanouthosts simply ignore the exceptions caught. This may hide wired issues and make debugging more difficult. We'd better let the issue exposed and be fixed. So, in this PR, I removed the try-except block.

How did you do it?
1. Removed the try-except block.
2. Skip processing for KVM DUTs, because they don't have fanout.

How did you verify/test it?
Verified on virtual testbed and physical testbeds.

Signed-off-by: Zhijian Li <zhijianli@microsoft.com>