[macsec]: Recover neighbor devices by sanity check  (#7449)
What is the motivation for this PR?
The neighbor devices may fail to initialize MACsec due to an abnormal condition that the MACsec interfaces weren't cleanupped in the last test.

How did you do it?
If "MACsec" is enabled to check whether the MACsec interface isn't in neighbor device. If not, recover the neighbor devices by config reload.

How did you verify/test it?
Check Azp

Signed-off-by: Ze Gan <ganze718@gmail.com>