[chassis] Fix dropcounter tests for T2 testbed with single and multi asic linecards (#7425)
What is the motivation for this PR?
In the current drop counter tests, for multi dut topologies it is assumed the port names will be same on all the duts.
This is not true on the device having a mix of single and multi asic linecards or linecard with different asics.
The PR fixes this.

How did you do it?
we drop counters are retrived for linecard then check if the port the tests is checking for existing on the dut before updating the drop counter list.

How did you verify/test it?
Run on the chassis with mix of single and multi asic linecards