Adding workaround fix for syslog issue found in test_sequential_restart.py (#6999)
What is the motivation for this PR?
To fix the issue reported in 6995.
In 202111 PTF T1 topo, after running platform_tests/test_sequential_restart.py with all BGP session established and routes populated, log messages are not getting written in to /var/log/syslog. The log messages are getting stuck. When we execute “systemctl stop syslog”, the stuck messages are getting written in to /var/log/syslog. This issue happens only when we had BGP session up and all the neighbors in T1 setup being up.

How did you do it?
Adding "sudo systemctl restart syslog" at the end of test to fix the issue.

How did you verify/test it?
Run the test and reported issue is not seen.