Add support of using lab tacacs server by default (#7624)
What is the motivation for this PR?
The current default configuration of DUTs does not support using lab tacacs server for authentication.
This change added a new variable tacacs_enabled_by_default in group_vars file to control enablement of lab tacacs server.

How did you do it?
The ansible playbook for deploy minigraph was updated to support configuring lab tacacs server based on value of this variable.
The tacacs test scripts are updated to support recover the lab tacacs server in a clean way after test is done.

How did you verify/test it?
Test run the tacacs scripts.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>