Exclude sku-sensors-data from the creds fixture (#7630)
What is the motivation for this PR?
The test_sensors.py script is using the creds fixture to get sku sensors data. This make the creds fixture floated because sku-sensors-data.yml is a big file. It does not make sense to include sensors data in a fixture for getting credentials.

How did you do it?
This change removed the sensors data from the creds fixture. For the test_sensors.py script, updated it to directly read sensors data from file sku-sensors-data.yml.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>