Merged PR 2026: Use the builtin service account to do merge and push
Use the builtin service account to do merge and push

After security setting of the sonic-mgmt-int repository is enhanced, the pipeline
for merging code change from github is broken. Previously the pipeline was using
a personal username and token for pushing code to sonic-mgmt-int. Although
the personal account is able to bypass the security policies configured for the
branches, push still fail.

This change updated the merge pipeline to use the default service account
to merge and push code.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>