[VM]Add IS-IS spf default interval testcase (#7659)
What is the motivation for this PR?
Add IS-IS spf default interval testcase

How did you do it?
Config spf interval in dut and update is-is metric in neighbor device, which will trigger lsp update. And check the spf run count in dut.

How did you verify/test it?
Run spf testcase and testcase would pass.