Fix flaky tacacs authorization one server down case
The tacacs authorization test has a case to verify that authorization
is still working if one of the tacacs servers is down.

In case the previous authorization method is local, after it is switched
to tacacs+, the tacacs client will start to contact the tacacs servers
for authorization. If we immediately run a command after the authorization
method is changed, this command may fail with authorization because the
client is still trying the invalid tacacs server.

The fix is to add a delay which is longer than the configured tacacs timeout
after authorization method is changed from local to tacacs+.

Extra improvements:
* Fixed issues detected by pre-commit.
* Created new test fixtures to configure and restore tacacs authorization
  configuration.
* Improved the code for starting and stopping tacacs server. The existing
  method of stopping tacacs server takes up to 40 seconds. Overall testing
  time of test_authorization.py takes up to 270 seconds. After the
  improvement, overall test time of this script is around 70 seconds.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>