Correct the condition mark topology for test_acl.py
What is the motivation for this PR?
Fix 5580, this test file is not support to run on ptf32.

How did you do it?
Changing the condition mark topology form "any" to ("t0", "t1", "t2", "m0", "mx")

How did you verify/test it?
N/A

Any platform specific information?
N/A

Supported testbed topology if it's a new test case?
N/A