[chassis][voq] remove conditional mark for voq tests (#6952)
What is the motivation for this PR?
The PR #5708 causes the tests in the files test_voq_nbr.py and test_voq_ipfwd.py to be skipped.

How did you do it?
This PR removes the voq test from condional mark and revert the the changes done in PR for these 2 files in PR #5708

How did you verify/test it?
Run voq tests

Signed-off-by: Arvindsrinivasan Lakshminarasimhan <arlakshm@microsoft.com>