[log-analyzer] Check both syslog and syslog.1 files when looking for marker (#7511)
What is the motivation for this PR?
In the current logic it waits for marker to appear in syslog or syslog.1 file.
To decide to jump from syslog to syslog.1 file, the logic checks ctime of the syslog file.

This is incorrect as ctime cannot be relied upon - https://unix.stackexchange.com/a/20464

This logic particularly creates problem when device boots up with no syslog.1 file.
This is often a case when logs are mounted on tmpfs partition, and old logs disappear post reboot.

This PR checks both syslog and syslog.1 files when looking for markers.

This fixes errors like below. NameError: global name 'FileNotFoundError' is due to FileNotFoundError not present in 202012 branch (Python2). That will be fixed separately and specifically in 202012 branch later.