Use enum_rand_one_per_hwsku_hostnameso that it has the correct enum_frontend_asic_index for the test_restart_swss/syncd tests (#7760)
platform_tests/test_sequential_restart.py::test_restart_swss is failing on chassis that has mixed LCs where the LCs has different number of front-end-asics.
It was found that by passing "rand_one_dut_hostname" as the paramter, it sometimes results to using the var of a LC that has more asics then the one chosen to perform the test that resulted to attempting to restart_swss for an asic that the LC does not have. We should use "enum_rand_one_per_hwsku_hostname" so that the proper front-end-asic is chosen correctly based on the chosen LC.