Update README.testbed.Setup.md (#7161)
What is the motivation for this PR?
The old setup document seems not clearer and easier for users. Also, KVM based VMs are facing shortcomings in front of cEOS like more consumption of both memory and interaction with the kernel. And docker-ptf need to be downloaded through internet resources, the local registry should be a better choice.

How did you do it?
Rewrite the logic and introduce cEOS clearer.

How did you verify/test it?
Follow the steps and deploy t0 topology based on cEOS and physical testbed.