[macsec]: Fix inserting none packet (#7787)
Approach
What is the motivation for this PR?
If to decap macsec failed, the packet with empty content will be inserted into the recent packet list. The packet with empty content cannot be parsed by scapy.

How did you do it?
If to decap macsec failed, to insert the original packet into the recent packet list.

How did you verify/test it?
Check Azp.

Signed-off-by: Ze Gan <ganze718@gmail.com>