Add support for t0-56-po2vlan topo to qos sai test (#7703)
What is the motivation for this PR?
Support t0-56-po2vlan topology in Qos SAI test
Currently the Qos SAI test doesn't support the topology t0-56-po2vlan, in this topology there are 2 downlink Vlans and a extra portchannel is the member of both vlans. Need to adjust the test to support this topology.

How did you do it?
The test will fail in __assignTestPortIps() in tests/qos/qos_sai_base.py because it doesn't expect there are multiple Vlans or a Vlan member is portchannel.
I add logic that when the topo is t0-56-po2vlan, choose the Vlan with physical ports to test and remove the portchannel member from the testVlanMembers.

How did you verify/test it?
By automation, all cases in test test_qos_sai.py passed on t0-56-po2vlan topo.