Fix issue in testPfcStormWithSharedHeadroomOccupancy (#7788)
- How did you do it?
We only reconfigure PFC WD polling timer when it is less than the existing value. When we recover the polling timer, we are enlarging it, which causes an error because we make the polling timer less than the detection timer, and restoration timer which are the same value as the current polling timer.

Fix
Restore, which is to enlarge, the detection timer and the restoration timer first and then the polling timer. The previous detection timer and the restoration timer are the same as the polling timer to be recovered.

- How did you verify/test it?
Run regression test

Signed-off-by: Stephen Sun <stephens@nvidia.com>