Merged PR 2216: enhance the logic to compare the routes
Why I did it:
When compare the routes directly, it could failed due to the mismatch of as path info.

How I did it:
We don't care the detailed path info, we only care about the route key, so only compare the keys would be good enough.

How I verified it:
Manually run the test