Replace scapy sniffer by tcpdump to overcome scapy performance issues when dealing with high load (#7707)
Scapy has a known performance limitation when it comes to high number of packets and high rates:
https://github.com/secdev/scapy/issues/2608

As a result, from time to time, packet losses are observed during fast and warm boot test.
For example, the pcap file would contain 40k packets while actually 45k packets were transmitted - the latest 5k packets will be missing from the pcap.

To overcome that, scapy sniffer was replaced with tcpdump which has much better performance.
Tcpdump is done on all the data ports in parallel.

Signed-off-by: Petro Pikh <petrop@nvidia.com>
Signed-off-by: yenlu <yenlu@cisco.com>