Fix router mac in the qos_sai test cases. (#7716)
In qos_sai testcases, currently the router_mac is obtained from the platform level command, which is not correct for multi-asic platforms. These platforms have different mac address based on asic. We need to update the script to use the correct function to get the correct mac address for the incoming interface.