Use thermals specified in SKIP_MODULES in inventory to calculate the expected number of thermals
test_thermals in  platform_tests/api/test_chassis.py fails on supervisor card in a chassis in which
not all fabric cards are present.

In platform.json file, we define a list of thermals present. For a supervisor card, the thermals
includes thermal info for the fabric cards as well. However, this info is for all the fabric cards
present in the chassis.

test_thermals uses this list of thermals in platform.json as the expected number of thermals present.
If we do not have all the fabric cards populated in the chassis, then the thermal data for the absent
fabric cards is missing. Thus, the test fails.

To fix the tests, we can define the thermals associated with fabric cards that are not present
under 'thermals' under 'skip_modules' for the supervisor card  in the inventory file

Example:

  ixre-cpm-chassis10:
      skip_modules:
        line-cards:
        fabric-cards:
          - FABRIC-CARD5
        psus:
        thermals:
          - sfm6_1(fan)
          - sfm6_2
          - sfm6_3
          - sfm6_4(fan
          - sfm6_5(fan)

We can then use this data to calculate the expected number of thermals