Change SONiC neighbor topo config to use exabgp and only config_db
For T0 topology with SONiC neighbor, the T1 peers don't connect to
exabgp on the PTF server and use a mix of config_db and manual FRR
configuration. (For T2, it does look like it connects to exabgp.) This
results in common SONiC commands such as `show ip bgp summary` not
working because of the manual FRR configuration, along with the config
not being saved in a single file. In addition, because of the lack of
connection to exabgp, prefixes are not advertised, and many test cases
that involve IO likely can't be tested.

Fix this by specifying all of the configuration within config_db, along
with adding configuration to connect to exabgp. Also, remove the manual
frr/zebra configuration files.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>