Fix "logging not defined" issue of some snappi tests (#7855)
What is the motivation for this PR?
The logging module is used by some files under tests.snappi. However, the module is not imported in the files. This caused error like below while collecting the snappi tests:
E NameError: name 'logging' is not defined

How did you do it?
The fix is to import logging in those files.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>